Rename __core__ to ap_map_flow.py
