+++ ApertureMapModelTools/__init__.py
@@ -11,10 +11,10 @@ import logging.config as _logging_config
 import os as _os
 import sys as _sys
 import PIL as _pil
-from .__core__ import DataField, FractureImageStack
-from .__core__ import _get_logger, set_main_logger_level
-from .__core__ import files_from_directory, load_infile_list
-from .__core__ import calc_percentile, calc_percentile_num, get_data_vect
+from .ap_map_flow import DataField, FractureImageStack
+from .ap_map_flow import _get_logger, set_main_logger_level
+from .ap_map_flow import files_from_directory, load_infile_list
+from .ap_map_flow import calc_percentile, calc_percentile_num, get_data_vect
 from . import data_processing
 from . import run_model
 from . import openfoam
